grading-server/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
scikit-image
scipy
Pillow
httpx
python-dotenv
python-multipart
